Implement cleanup command
This is new command that can be used to delete everything running in a
cluster related to an application using the labels we use.

This currently uses a hardcoded set of resource types. It would be
possible to make this generic, which might become needed in the future.
See comments in the code.

---

The overall design for how to use the command:

```sh
tye cleanup
```

Will read a solution or tye.yaml to get the application name. It then
enumerates and deletes all resources in the cluser with
`app.kubernetes.io/part-of=application-name` (in the current namespace).

```sh
tye cleanup --what-if
```

Will do the same, but without actually doing the deletion.

```sh
tye cleanup -i
```

Will interactively give you a yes/no choice for whether to delete each
resource.